test_8

+++ 2D/training/train.py
@@ -57,7 +57,7 @@ for epoch in range (num_epochs):
     for batch_idx, batch in enumerate(dataloader):
         data = batch[:, np.newaxis, :, :]
         optimizer.zero_grad()
-        data = torch.DoubleTensor(data) 
+        data = torch.FloatTensor(data) 
         data_norm = torch.ones((data.size(0), data.size(1), 1, 1)) / ratio_max
         output = model(data)
         # loss = laplacian_loss(output, data = data, data_norm = data_norm)
